Speed up square root function by applying john carmack wizard-magic

+++ engine/class_modules/sc_enemy.cpp
@@ -1137,7 +1137,7 @@ void enemy_t::init_base_stats()
 
   if ( sim -> overrides.target_health.size() > 0 )
   {
-    initial_health = sim -> overrides.target_health[ enemy_id % sim -> overrides.target_health.size() ];
+    initial_health = static_cast<double>( sim -> overrides.target_health[ enemy_id % sim -> overrides.target_health.size() ] );
   }
   else
   {
